+++ module-common/src/main/java/com/welcommu/modulecommon/exception/GlobalControllerAdvice.java
@@ -1,6 +1,6 @@
 package com.welcommu.modulecommon.exception;
 
-import com.welcommu.modulecommon.exception.dto.ErrorResponse;
+import com.welcommu.modulecommon.dto.ErrorResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
