+++ module-common/src/main/resources/application.properties
@@ -14,4 +14,4 @@ cors.allowedMethods=GET,POST,PUT,DELETE,OPTIONS
 cors.allowedHeaders=*
 cors.allowCredentials=true
 logging.level.com.welcommu=DEBUG
-logging.level.com.welcommu.modulecommon.token.helper.JwtTokenHelper=DEBUG
+logging.level.com.welcommu.modulecommon.auth.token.JwtTokenHelper=DEBUG
